+++ backend/packages/kernel/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cat-factory/kernel
 
+## 0.342.1
+
+### Patch Changes
+
+- Updated dependencies [6ff632f]
+  - @cat-factory/contracts@0.350.0
+
 ## 0.342.0
 
 ### Minor Changes